+++ ci/release_notes.md
@@ -1,1 +1,3 @@
-Empty - please add release notes here
+## Bug Fixes
+
+- Fix `Error: Couldn't find environment` after adding new environment to `cepler.yml`